fix(ci): move me-* test files out of api/functions/ root
Netlify scans every .ts file in the functions directory and bundles each as
a separate function. With `api/functions/me-password.test.ts` sitting next
to `api/functions/me-password.ts`, the bundler sees two candidates for the
`me-password` function slot. The deploy preview ended up with only
`me-settings` deployed (401 = handler ran but unauthorized) and
`me-username` / `me-password` returning 404 (the .test.ts file won the
slot — no exported handler → 404).

Repo convention: no co-located tests in api/functions/ (admin-verify and
other functions put handler-level tests in apps/web/tests/unit/). Move the
three new .test.ts files into api/functions/__tests__/ — Netlify ignores
subdirectories by default, so they no longer conflict. vitest's include
pattern (`api/functions/**/*.test.ts`) picks them up at the new location.
Relative imports updated: `./me-password` → `../me-password`, same for
`./db` and `./ratelimit` vi.mock calls.

Caught by deploy-preview failure on PR #294 (UNS-31 PR 1) — Header/Redirect
rules checks fail because Netlify's post-build validation rejects the
malformed function inventory.
